# `AffineConstraints`: Improve loop