closes https://github.com/dealii/dealii/issues/17692