@marcfehling 2024-09-13
This is a lot faster now!

Before this patch:
```
$ mpirun -np 2 ./merge-constraints 
Set up grid and dofs.
  Number of cells: 391022
  Number of DoFs:  33414037
Make hanging node constraints.
Make constraints consistent.

+---------------------------------------------------------+------------------+------------+------------------+
| Total wallclock time elapsed                            |      1603s     0 |      1603s |      1603s     1 |
|                                                         |                  |                               |
| Section                                     | no. calls |   min time  rank |   avg time |   max time  rank |
+---------------------------------------------------------+------------------+------------+------------------+
| run                                         |         1 |      1603s     0 |      1603s |      1603s     1 |
|  set_up                                     |         1 |      1.77s     0 |      1.77s |      1.77s     1 |
|  make_hanging_node_constraints              |         1 |     1.399s     0 |     1.713s |     2.028s     1 |
|  make_consistent_in_parallel                |         1 |      1599s     1 |      1599s |      1600s     0 |
|   1) compute_locally_relevant_constraints   |         2 |      1583s     1 |      1583s |      1584s     0 |
|      step 0                                 |         2 |   0.02087s     0 |   0.03148s |   0.04209s     1 |
|      step 1                                 |         2 |   0.04238s     1 |     1.301s |     2.559s     0 |
|      step 2                                 |         2 |     1.633s     0 |     1.793s |     1.953s     1 |
|      step 3                                 |         2 |      1580s     0 |      1580s |      1581s     1 |
|      step 3: determine owners               |         2 |   0.03986s     1 |    0.1999s |    0.3598s     0 |
|      step 3: loop rank_and_indices          |         2 |     81.98s     1 |     830.3s |      1579s     0 |
|      step 3: loop rank_and_indices: find_if |    276716 |     81.95s     1 |     829.8s |      1578s     0 | <---
|      step 3: some_to_some                   |         2 |   0.01266s     0 |     748.3s |      1497s     1 |
|      step 3: collate                        |         2 |    0.5216s     0 |     1.312s |     2.102s     1 |
|   2)                                        |         2 |     5.492s     0 |     6.687s |     7.883s     1 |
|   3)                                        |         2 |     2.265s     0 |     2.739s |     3.212s     1 |
|   4)                                        |         2 | 7.466e-05s     1 |     2.459s |     4.918s     0 |
+---------------------------------------------------------+------------------+------------+------------------+
```

After this patch:
```
+----------------------------------------------------------+------------------+------------+------------------+
| Total wallclock time elapsed                             |     24.89s     0 |      24.9s |     24.92s     1 |
|                                                          |                  |                               |
| Section                                      | no. calls |   min time  rank |   avg time |   max time  rank |
+----------------------------------------------------------+------------------+------------+------------------+
| run                                          |         1 |     24.88s     0 |      24.9s |     24.92s     1 |
|  set_up                                      |         1 |     1.804s     0 |     1.804s |     1.804s     1 |
|  make_hanging_node_constraints               |         1 |     1.441s     0 |     1.758s |     2.076s     1 |
|  make_consistent_in_parallel                 |         1 |     20.89s     1 |      21.2s |     21.51s     0 |
|   1) compute_locally_relevant_constraints    |         2 |     4.427s     1 |     4.838s |     5.249s     0 |
|      step 0                                  |         2 |   0.02313s     0 |   0.03386s |   0.04458s     1 |
|      step 1                                  |         2 |   0.04478s     1 |       1.3s |     2.556s     0 |
|      step 2                                  |         2 |      1.72s     0 |     1.868s |     2.015s     1 |
|      step 3                                  |         2 |    0.9499s     0 |     1.636s |     2.321s     1 |
|      step 3: determine owners                |         2 |   0.04119s     1 |    0.1889s |    0.3365s     0 |
|      step 3: loop rank_and_indices           |         2 |     0.024s     1 |   0.02474s |   0.02549s     0 |
|      step 3: loop rank_and_indices: new loop |         2 |   0.02398s     1 |   0.02473s |   0.02547s     0 | <---
|      step 3: some_to_some                    |         2 |   0.01198s     0 |   0.01273s |   0.01348s     1 |
|      step 3: collate                         |         2 |    0.5707s     0 |     1.404s |     2.238s     1 |
|   2)                                         |         2 |     5.563s     0 |     6.755s |     7.947s     1 |
|   3)                                         |         2 |     2.346s     0 |     2.841s |     3.336s     1 |
|   4)                                         |         2 | 6.195e-05s     1 |     2.521s |     5.041s     0 |
+----------------------------------------------------------+------------------+------------+------------------+
```